# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "媒资管理系统 REST API。成功响应直接返回业务 JSON（camelCase）；失败返回 { err_code, err_msg }。",
        "title": "GIN-MAM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/asset-workflow/all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "全部审核实例",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.PageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/assigned-to-me": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "待我审核列表",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.PageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/audit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "对单个流程实例执行通过或打回",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "审核实例",
                "parameters": [
                    {
                        "description": "审核参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.auditRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/audited-by-me": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "我审过的",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.PageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/created-by-me": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "我发起的审核",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.PageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/multi-audit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "批量审核",
                "parameters": [
                    {
                        "description": "批量审核参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.multiAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/revoke": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "撤回本人发起的未完成审核实例",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "撤回审核",
                "parameters": [
                    {
                        "description": "实例 ID 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.revokeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/asset-workflow/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将媒资提交至指定流程定义，创建审核实例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流"
                ],
                "summary": "提交审核",
                "parameters": [
                    {
                        "description": "提交参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.submitRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.InstanceVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
...
        "/api/v1/workflow/def": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流定义"
                ],
                "summary": "创建流程定义",
                "parameters": [
                    {
                        "description": "流程定义",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.createDefRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.DefVO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        },
        "/api/v1/workflow/def/page": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作流定义"
                ],
                "summary": "流程定义分页",
                "parameters": [
                    {
                        "description": "分页参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.pageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_workflow.DefPageResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_gin-mam_backend_internal_pkg_httpx.ErrorVo": {
            "type": "object",
            "properties": {
                "err_code": {
                    "type": "integer",
                    "example": 40100
                },
                "err_msg": {
                    "type": "string",
                    "example": "未授权"
                }
            }
        },
...
        "internal_domain_workflow.DefLevelUserVO": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.DefPageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_workflow.DefVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_workflow.DefVO": {
            "type": "object",
            "properties": {
                "auditLevel": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "levelUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_workflow.DefLevelUserVO"
                    }
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.InstanceVO": {
            "type": "object",
            "properties": {
                "assetId": {
                    "type": "string"
                },
                "auditLevel": {
                    "type": "integer"
                },
                "auditStatus": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workflowDefId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.PageResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_workflow.InstanceVO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_domain_workflow.auditItem": {
            "type": "object",
            "required": [
                "instanceId"
            ],
            "properties": {
                "instanceId": {
                    "type": "string"
                },
                "pass": {
                    "type": "boolean"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.auditRequest": {
            "type": "object",
            "required": [
                "instanceId"
            ],
            "properties": {
                "instanceId": {
                    "type": "string"
                },
                "pass": {
                    "type": "boolean"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.createDefRequest": {
            "type": "object",
            "required": [
                "auditLevel",
                "name"
            ],
            "properties": {
                "auditLevel": {
                    "type": "integer"
                },
                "levelUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_workflow.levelUserItem"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.levelUserItem": {
            "type": "object",
            "required": [
                "level",
                "userId"
            ],
            "properties": {
                "level": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "internal_domain_workflow.multiAuditRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domain_workflow.auditItem"
                    }
                }
            }
        },
        "internal_domain_workflow.pageRequest": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "internal_domain_workflow.revokeRequest": {
            "type": "object",
            "required": [
                "instanceIds"
            ],
            "properties": {
                "instanceIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_domain_workflow.submitRequest": {
            "type": "object",
            "required": [
                "assetId",
                "workflowDefId"
            ],
            "properties": {
                "assetId": {
                    "type": "string"
                },
                "workflowDefId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}